requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.4
django-jazzmin==3.0.1
djangorestframework==3.16.1
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
